pyproject.toml
[project]
name = "nyc-taxi-data-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "delta-spark==4.0.0",
    "pyspark==4.0.0",
]
